[Feat:#13] - home 컴포넌트 로직 분리
